+++ containers/SmallVector.h
@@ -31,19 +31,19 @@ namespace evo{
 			// constructors / destructors
 
 
-			EVO_NODISCARD SmallVector() noexcept = default;
-			~SmallVector() noexcept = default;
+			EVO_NODISCARD inline SmallVector() noexcept = default;
+			inline ~SmallVector() noexcept = default;
 
 
 			///////////////////////////////////
 			// copy
 
-			EVO_NODISCARD SmallVector(const SmallVector<T, SMALL_CAPACITY>& rhs) noexcept {
+			EVO_NODISCARD inline SmallVector(const SmallVector<T, SMALL_CAPACITY>& rhs) noexcept {
 				this->data_block = rhs.data_block;
 			};
 
 
-			auto operator=(const SmallVector<T, SMALL_CAPACITY>& rhs) noexcept -> SmallVector<T, SMALL_CAPACITY>& {
+			inline auto operator=(const SmallVector<T, SMALL_CAPACITY>& rhs) noexcept -> SmallVector<T, SMALL_CAPACITY>& {
 				this->data_block = rhs.data_block;
 
 				return *this;
@@ -53,12 +53,12 @@ namespace evo{
 			///////////////////////////////////
 			// move
 
-			EVO_NODISCARD SmallVector(SmallVector<T, SMALL_CAPACITY>&& rhs) noexcept {
+			EVO_NODISCARD inline SmallVector(SmallVector<T, SMALL_CAPACITY>&& rhs) noexcept {
 				this->data_block = std::move(rhs.data_block);
 			};
 
 
-			auto operator=(SmallVector<T, SMALL_CAPACITY>&& rhs) noexcept -> SmallVector<T, SMALL_CAPACITY>& {
+			inline auto operator=(SmallVector<T, SMALL_CAPACITY>&& rhs) noexcept -> SmallVector<T, SMALL_CAPACITY>& {
 				this->data_block = std::move(rhs.data_block);
 
 				return *this;
@@ -75,7 +75,7 @@ namespace evo{
 			///////////////////////////////////
 			// construct from initializer list
 
-			EVO_NODISCARD SmallVector(std::initializer_list<T> init_list) noexcept {
+			EVO_NODISCARD inline SmallVector(std::initializer_list<T> init_list) noexcept {
 				this->clear();
 
 				if(init_list.size() <= SMALL_CAPACITY){
@@ -85,7 +85,7 @@ namespace evo{
 				}
 			};
 
-			auto operator=(std::initializer_list<T> init_list) noexcept -> SmallVector<T, SMALL_CAPACITY>& {
+			inline auto operator=(std::initializer_list<T> init_list) noexcept -> SmallVector<T, SMALL_CAPACITY>& {
 				this->clear();
 
 				if(init_list.size() <= SMALL_CAPACITY){
@@ -107,7 +107,7 @@ namespace evo{
 			///////////////////////////////////
 			// at
 
-			EVO_NODISCARD auto at(size_t index) noexcept -> T& {
+			EVO_NODISCARD inline auto at(size_t index) noexcept -> T& {
 				EVO_DEBUG_ASSERT(index < this->size());
 
 				if(this->is_small()){
@@ -117,7 +117,7 @@ namespace evo{
 				}
 			};
 
-			EVO_NODISCARD auto at(size_t index) const noexcept -> const T& {
+			EVO_NODISCARD inline auto at(size_t index) const noexcept -> const T& {
 				EVO_DEBUG_ASSERT(index < this->size());
 
 				if(this->is_small()){
@@ -129,30 +129,30 @@ namespace evo{
 
 
 			template<std::enable_if<std::is_same_v<size_type, size_t>> = false>
-			EVO_NODISCARD auto at(size_type index) noexcept -> T& { return this->at(size_t(index)); };
+			EVO_NODISCARD inline auto at(size_type index) noexcept -> T& { return this->at(size_t(index)); };
 
 			template<std::enable_if<std::is_same_v<size_type, size_t>> = false>
-			EVO_NODISCARD auto at(size_type index) const noexcept -> const T& { return this->at(size_t(index)); };
+			EVO_NODISCARD inline auto at(size_type index) const noexcept -> const T& { return this->at(size_t(index)); };
 
 
 
 			///////////////////////////////////
 			// operator[]
 
-			EVO_NODISCARD auto operator[](size_t index)       noexcept ->       T& { return this->at(index); };
-			EVO_NODISCARD auto operator[](size_t index) const noexcept -> const T& { return this->at(index); };
+			EVO_NODISCARD inline auto operator[](size_t index)       noexcept ->       T& { return this->at(index); };
+			EVO_NODISCARD inline auto operator[](size_t index) const noexcept -> const T& { return this->at(index); };
 
 			template<std::enable_if<std::is_same_v<size_type, size_t>> = false>
-			EVO_NODISCARD auto operator[](size_type index)       noexcept ->       T& { return this->at(index); };
+			EVO_NODISCARD inline auto operator[](size_type index)       noexcept ->       T& { return this->at(index); };
 
 			template<std::enable_if<std::is_same_v<size_type, size_t>> = false>
-			EVO_NODISCARD auto operator[](size_type index) const noexcept -> const T& { return this->at(index); };
+			EVO_NODISCARD inline auto operator[](size_type index) const noexcept -> const T& { return this->at(index); };
 
 
 			///////////////////////////////////
 			// front
 
-			EVO_NODISCARD auto front() noexcept -> T& {
+			EVO_NODISCARD inline auto front() noexcept -> T& {
 				if(this->is_small()){
 					return this->get_small().front();
 				}else{
@@ -160,7 +160,7 @@ namespace evo{
 				}
 			};
 
-			EVO_NODISCARD auto front() const noexcept -> const T& {
+			EVO_NODISCARD inline auto front() const noexcept -> const T& {
 				if(this->is_small()){
 					return this->get_small().front();
 				}else{
@@ -172,7 +172,7 @@ namespace evo{
 			///////////////////////////////////
 			// back
 
-			EVO_NODISCARD auto back() noexcept -> T& {
+			EVO_NODISCARD inline auto back() noexcept -> T& {
 				if(this->is_small()){
 					return this->get_small().back();
 				}else{
@@ -180,7 +180,7 @@ namespace evo{
 				}
 			};
 
-			EVO_NODISCARD auto back() const noexcept -> const T& {
+			EVO_NODISCARD inline auto back() const noexcept -> const T& {
 				if(this->is_small()){
 					return this->get_small().back();
 				}else{
@@ -192,7 +192,7 @@ namespace evo{
 			///////////////////////////////////
 			// data
 
-			EVO_NODISCARD auto data() noexcept -> T* {
+			EVO_NODISCARD inline auto data() noexcept -> T* {
 				if(this->is_small()){
 					return this->get_small().data();
 				}else{
@@ -200,7 +200,7 @@ namespace evo{
 				}
 			};
 
-			EVO_NODISCARD auto data() const noexcept -> const T* {
+			EVO_NODISCARD inline auto data() const noexcept -> const T* {
 				if(this->is_small()){
 					return this->get_small().data();
 				}else{
@@ -212,29 +212,29 @@ namespace evo{
 			//////////////////////////////////////////////////////////////////////
 			// iterators
 
-			EVO_NODISCARD auto begin()        noexcept -> iterator       { return iterator{this->data()};       };
-			EVO_NODISCARD auto begin()  const noexcept -> const_iterator { return const_iterator{this->data()}; };
-			EVO_NODISCARD auto cbegin() const noexcept -> const_iterator { return const_iterator{this->data()}; };
+			EVO_NODISCARD inline auto begin()        noexcept -> iterator       { return iterator{this->data()};       };
+			EVO_NODISCARD inline auto begin()  const noexcept -> const_iterator { return const_iterator{this->data()}; };
+			EVO_NODISCARD inline auto cbegin() const noexcept -> const_iterator { return const_iterator{this->data()}; };
 
-			EVO_NODISCARD auto end()        noexcept -> iterator       { return iterator{&this->data()[this->size()]};       };
-			EVO_NODISCARD auto end()  const noexcept -> const_iterator { return const_iterator{&this->data()[this->size()]}; };
-			EVO_NODISCARD auto cend() const noexcept -> const_iterator { return const_iterator{&this->data()[this->size()]}; };
+			EVO_NODISCARD inline auto end()        noexcept -> iterator       { return iterator{&this->data()[this->size()]};       };
+			EVO_NODISCARD inline auto end()  const noexcept -> const_iterator { return const_iterator{&this->data()[this->size()]}; };
+			EVO_NODISCARD inline auto cend() const noexcept -> const_iterator { return const_iterator{&this->data()[this->size()]}; };
 
 
-			EVO_NODISCARD auto rbegin()        noexcept -> reverse_iterator       {return reverse_iterator{&this->data()[this->size()-1]};      };
-			EVO_NODISCARD auto rbegin()  const noexcept -> const_reverse_iterator {return const_reverse_iterator{&this->data()[this->size()-1]};};
-			EVO_NODISCARD auto crbegin() const noexcept -> const_reverse_iterator {return const_reverse_iterator{&this->data()[this->size()-1]};};
+			EVO_NODISCARD inline auto rbegin()        noexcept -> reverse_iterator       {return reverse_iterator{&this->data()[this->size()-1]};      };
+			EVO_NODISCARD inline auto rbegin()  const noexcept -> const_reverse_iterator {return const_reverse_iterator{&this->data()[this->size()-1]};};
+			EVO_NODISCARD inline auto crbegin() const noexcept -> const_reverse_iterator {return const_reverse_iterator{&this->data()[this->size()-1]};};
 
-			EVO_NODISCARD auto rend()        noexcept -> reverse_iterator       { return reverse_iterator{this->data()[-1]};       };
-			EVO_NODISCARD auto rend()  const noexcept -> const_reverse_iterator { return const_reverse_iterator{this->data()[-1]}; };
-			EVO_NODISCARD auto crend() const noexcept -> const_reverse_iterator { return const_reverse_iterator{this->data()[-1]}; };
+			EVO_NODISCARD inline auto rend()        noexcept -> reverse_iterator       { return reverse_iterator{this->data()[-1]};       };
+			EVO_NODISCARD inline auto rend()  const noexcept -> const_reverse_iterator { return const_reverse_iterator{this->data()[-1]}; };
+			EVO_NODISCARD inline auto crend() const noexcept -> const_reverse_iterator { return const_reverse_iterator{this->data()[-1]}; };
 
 
 
 			//////////////////////////////////////////////////////////////////////
 			// capacity
 
-			EVO_NODISCARD auto empty() const noexcept -> bool {
+			EVO_NODISCARD inline auto empty() const noexcept -> bool {
 				if(this->is_small()){
 					return this->get_small().empty();
 				}else{
@@ -242,7 +242,7 @@ namespace evo{
 				}
 			};
 
-			EVO_NODISCARD auto size() const noexcept -> size_t {
+			EVO_NODISCARD inline auto size() const noexcept -> size_t {
 				if(this->is_small()){
 					return this->get_small().size();
 				}else{
@@ -254,7 +254,7 @@ namespace evo{
 				return size_t(std::numeric_limits<difference_type>::max() / sizeof(T));
 			};
 
-			EVO_NODISCARD auto capacity() const noexcept -> size_t {
+			EVO_NODISCARD inline auto capacity() const noexcept -> size_t {
 				if(this->is_small()){
 					return this->get_small().capactity();
 				}else{
@@ -277,8 +277,10 @@ namespace evo{
 			///////////////////////////////////
 			// insert
 
-			auto insert(const_iterator pos, const T& value) noexcept -> iterator {
-				EVO_DEBUG_ASSERT(this->current_size < this->capacity());
+			inline auto insert(const_iterator pos, const T& value) noexcept -> iterator {
+				if(this->is_small() && this->size() == SMALL_CAPACITY){
+					this->copy_to_big();
+				}
 
 				*this->end() = value;
 
@@ -296,8 +298,10 @@ namespace evo{
 			};
 
 
-			auto insert(const_iterator pos, T&& value) noexcept -> iterator {
-				EVO_DEBUG_ASSERT(this->current_size < this->capacity());
+			inline auto insert(const_iterator pos, T&& value) noexcept -> iterator {
+				if(this->is_small() && this->size() == SMALL_CAPACITY){
+					this->copy_to_big();
+				}
 
 				*this->end() = std::move(value);
 
@@ -327,8 +331,10 @@ namespace evo{
 			// emplace
 
 			template<class... Args>
-			auto emplace(const_iterator pos, Args&&... args) noexcept -> iterator {
-				EVO_DEBUG_ASSERT(this->current_size < this->capacity());
+			inline auto emplace(const_iterator pos, Args&&... args) noexcept -> iterator {
+				if(this->is_small() && this->size() == SMALL_CAPACITY){
+					this->copy_to_big();
+				}
 
 				std::construct_at(&*this->end(), std::forward<Args>(args)...);
 
@@ -349,7 +355,7 @@ namespace evo{
 			///////////////////////////////////
 			// erase
 
-			auto erase(const_iterator pos) noexcept -> iterator {
+			inline auto erase(const_iterator pos) noexcept -> iterator {
 				std::destroy_at(&*pos);
 
 				iterator output = [&](){
@@ -373,23 +379,29 @@ namespace evo{
 			///////////////////////////////////
 			// push_back
 
-			auto push_back(const T& value) noexcept -> void {
-				EVO_DEBUG_ASSERT(this->size() < SMALL_CAPACITY);
-
+			inline auto push_back(const T& value) noexcept -> void {
 				if(this->is_small()){
-					return this->get_small().push_back(value);
+					if(this->size() == SMALL_CAPACITY){
+						this->copy_to_big();
+						return this->get_big().push_back(value);
+					}else{
+						return this->get_small().push_back(value);
+					}
 				}else{
 					return this->get_big().push_back(value);
 				}
 			};
 
-			auto push_back(T&& value) noexcept -> void {
-				EVO_DEBUG_ASSERT(this->size() < SMALL_CAPACITY);
-
+			inline auto push_back(T&& value) noexcept -> void {
 				if(this->is_small()){
-					return this->get_small().push_back(std::move(value));
+					if(this->size() == SMALL_CAPACITY){
+						this->copy_to_big();
+						return this->get_big().push_back(value);
+					}else{
+						return this->get_small().push_back(value);
+					}
 				}else{
-					return this->get_big().push_back(std::move(value));
+					return this->get_big().push_back(value);
 				}
 			};
 
@@ -397,11 +409,14 @@ namespace evo{
 			///////////////////////////////////
 			// emplace_back
 
-			auto emplace_back(auto&&... args) noexcept -> T& {
-				EVO_DEBUG_ASSERT(this->size() < SMALL_CAPACITY);
-
+			inline auto emplace_back(auto&&... args) noexcept -> T& {
 				if(this->is_small()){
-					return this->get_small().emplace_back(std::forward<decltype(args)>(args)...);
+					if(this->size() == SMALL_CAPACITY){
+						this->copy_to_big();
+						return this->get_big().emplace_back(std::forward<decltype(args)>(args)...);
+					}else{
+						return this->get_small().emplace_back(std::forward<decltype(args)>(args)...);
+					}
 				}else{
 					return this->get_big().emplace_back(std::forward<decltype(args)>(args)...);
 				}
@@ -413,7 +428,7 @@ namespace evo{
 			///////////////////////////////////
 			// pop_back
 
-			auto pop_back() noexcept -> void {
+			inline auto pop_back() noexcept -> void {
 				EVO_DEBUG_ASSERT(this->size() > 0);
 
 				if(this->is_small()){
@@ -428,15 +443,18 @@ namespace evo{
 			// resize
 
 			template<std::enable_if<std::is_same_v<size_type, size_t>> = false>
-			auto resize(size_type new_size) noexcept -> void {
+			inline auto resize(size_type new_size) noexcept -> void {
 				this->resize(size_t(new_size));
 			};
 
-			auto resize(size_t new_size) noexcept -> void {
-				EVO_DEBUG_ASSERT(new_size < SMALL_CAPACITY);
-
+			inline auto resize(size_t new_size) noexcept -> void {
 				if(this->is_small()){
-					return this->get_small().resize(new_size);
+					if(new_size >= SMALL_CAPACITY){
+						this->copy_to_big();
+						return this->get_big().resize(new_size);
+					}else{
+						return this->get_small().resize(new_size);
+					}
 				}else{
 					return this->get_big().resize(new_size);
 				}
@@ -444,9 +462,6 @@ namespace evo{
 
 
 			// not implemented yet: swap
-
-
-
 
 
 
@@ -487,6 +502,20 @@ namespace evo{
 			inline auto emplace_data(auto&&... args) noexcept -> void {
 				// I have absolutely no idea why I need to write it this way, but it won't compile otherwise
 				this->data_block.template emplace<U>(std::forward<decltype(args)>(args)...);
+			};
+
+
+			inline auto copy_to_big() noexcept -> void {
+				auto temp_holder = std::vector<T>();
+				temp_holder.reserve(this->size());
+
+				for(T& elem : this->get_small()){
+					temp_holder.emplace_back(std::move(elem));
+				}
+
+				this->get_small().clear();
+
+				this->emplace_data<std::vector<T>>(std::move(temp_holder));
 			};